Remove host build rules for libffi, and no longer build it for x86 device targets.
The new recommendation for libffi usage is: Libffi is fine to use as a
means of bootstrapping your device build for a new architecture, but
you probably want to replace it with custom code before you ship. As
for a host build, if you need it (which isn't actually that likely),
you should use the libffi that comes with your platform or build it as
a normal library (e.g., download it and then "./configure; make
install").

Change-Id: I710daa136a2cacc4f95cdf6b5fd60a3d3c072df5