cheets: Fix the cheets-libffi MODULE_LICENSE type
The AOSP repo identifies the libffi license to be a BSD-like license,
and since I started with it (but updated to a more recent version), I
inherited the MODULE_LICENSE_BSD_LIKE file.

https://android.googlesource.com/platform/external/libffi/+/master

However the actual license type is MIT.

https://opensource.org/licenses/BSD-2-Clause
https://opensource.org/licenses/MIT

(compare with the LICENSE and NOTICE files)

I realized in checking over a change landed on the Chrome side where it
was added fresh that this needed to be fixed here.

This patch renames the file to MODULE_LICENSE_MIT to correctly set the
license that applies.

Change-Id: If59ebd66ba9c2eb1d7c3f9488eb3a42075334ebe