[nyc-arc] Fix cheets-libffi compilation on ARM
This is a bug in LLVM: https://llvm.org/bugs/show_bug.cgi?id=20595 . It
was marked as WONTFIX and later fixed upstream in libffi in a different
way via https://github.com/libffi/libffi/commit/e7f15f60e but it does
not cleanly apply. Instead, use the fix mentioned in the LLVM bug
thread.

Bug: 31833456
Test: lunch cheets_arm-eng && m
Change-Id: I602c32dd1463e8aeca95cd06dd19193dee008114