Support building as shared library and remove win32.S
It turns out that this win32.S file is not part of the upstream
repository, and it thus has duplicate definitions (as libffi has evolved
proper ones over time it seems). This allows libffi to also now build as
a shared library as well, so I went ahead and enabled that as well.

Bug: http://b/162610242
Test: mmma . (in external/libffi)
Change-Id: Iee7ea2435817718b626edb8a0582f37c7eb7d110